refactor: move instructions to general
