+++ frontend/src/general/AddEmbedInstructions.svelte
